api/requirements.txt
asgiref==3.8.1
boto3==1.36.21
botocore==1.36.21
chardet==5.2.0
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jmespath==1.0.1
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
reportlab==4.3.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
django-filter==25.1